+++ lib/src/directory_watcher/windows.dart
@@ -12,12 +12,16 @@ import 'dart:io';
 import 'package:path/path.dart' as p;
 
 import '../constructable_file_system_event.dart';
+import '../directory_watcher.dart';
 import '../path_set.dart';
+import '../resubscribable.dart';
 import '../utils.dart';
 import '../watch_event.dart';
-import 'resubscribable.dart';
 
-class WindowsDirectoryWatcher extends ResubscribableDirectoryWatcher {
+class WindowsDirectoryWatcher extends ResubscribableWatcher
+    implements DirectoryWatcher {
+  String get directory => path;
+
   WindowsDirectoryWatcher(String directory)
       : super(directory, () => new _WindowsDirectoryWatcher(directory));
 }
@@ -40,8 +44,10 @@ class _EventBatcher {
   }
 }
 
-class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
-  final String directory;
+class _WindowsDirectoryWatcher
+    implements DirectoryWatcher, ManuallyClosedWatcher {
+  String get directory => path;
+  final String path;
 
   Stream<WatchEvent> get events => _eventsController.stream;
   final _eventsController = new StreamController<WatchEvent>.broadcast();
@@ -79,8 +85,9 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
   final Set<StreamSubscription<FileSystemEntity>> _listSubscriptions
       = new HashSet<StreamSubscription<FileSystemEntity>>();
 
-  _WindowsDirectoryWatcher(String directory)
-      : directory = directory, _files = new PathSet(directory) {
+  _WindowsDirectoryWatcher(String path)
+      : path = path,
+        _files = new PathSet(path) {
     // Before we're ready to emit events, wait for [_listDir] to complete.
     _listDir().then((_) {
       _startWatch();
@@ -110,12 +117,12 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
   /// On Windows, if [directory] is deleted, we will not receive any event.
   ///
   /// Instead, we add a watcher on the parent folder (if any), that can notify
-  /// us about [directory]. This also includes events such as moves.
+  /// us about [path]. This also includes events such as moves.
   void _startParentWatcher() {
-    var absoluteDir = p.absolute(directory);
+    var absoluteDir = p.absolute(path);
     var parent = p.dirname(absoluteDir);
-    // Check if [directory] is already the root directory.
-    if (FileSystemEntity.identicalSync(parent, directory)) return;
+    // Check if [path] is already the root directory.
+    if (FileSystemEntity.identicalSync(parent, path)) return;
     var parentStream = new Directory(parent).watch(recursive: false);
     _parentWatchSubscription = parentStream.listen((event) {
       // Only look at events for 'directory'.
@@ -127,7 +134,7 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
       // the directory is now gone.
       if (event is FileSystemMoveEvent ||
           event is FileSystemDeleteEvent ||
-          (FileSystemEntity.typeSync(directory) ==
+          (FileSystemEntity.typeSync(path) ==
            FileSystemEntityType.NOT_FOUND)) {
         for (var path in _files.toSet()) {
           _emitEvent(ChangeType.REMOVE, path);
@@ -210,7 +217,7 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
   /// CREATE event for the destination.
   ///
   /// The returned events won't contain any [FileSystemMoveEvent]s, nor will it
-  /// contain any events relating to [directory].
+  /// contain any events relating to [path].
   Map<String, Set<FileSystemEvent>> _sortEvents(List<FileSystemEvent> batch) {
     var eventsForPaths = {};
 
@@ -360,7 +367,7 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
   /// Start or restart the underlying [Directory.watch] stream.
   void _startWatch() {
     // Batch the events together so that we can dedup events.
-    var innerStream = new Directory(directory).watch(recursive: true);
+    var innerStream = new Directory(path).watch(recursive: true);
     _watchSubscription = innerStream.listen(_onEvent,
         onError: _eventsController.addError,
         onDone: _onDone);
@@ -374,7 +381,7 @@ class _WindowsDirectoryWatcher implements ManuallyClosedDirectoryWatcher {
 
     _files.clear();
     var completer = new Completer();
-    var stream = new Directory(directory).list(recursive: true);
+    var stream = new Directory(path).list(recursive: true);
     void handleEntity(entity) {
       if (entity is! Directory) _files.add(entity.path);
     }