+++ test/directory_watcher/linux_test.dart
@@ -27,7 +27,7 @@ void main() {
       () {
     withPermutations((i, j, k) =>
         writeFile("dir/sub/sub-$i/sub-$j/file-$k.txt"));
-    startWatcher(dir: "dir");
+    startWatcher(path: "dir");
 
     renameDir("dir/sub", "sub");
     renameDir("sub", "dir/sub");